+++ language/system/OtOS.cpp
@@ -271,8 +271,8 @@ OtObject OtOSClass::networks() {
 //
 
 double OtOSClass::clock() {
-	return (double) std::chrono::duration_cast<std::chrono::milliseconds>(
-		std::chrono::system_clock::now().time_since_epoch()).count() / 1000.0;
+	return static_cast<float>(std::chrono::duration_cast<std::chrono::milliseconds>(
+		std::chrono::system_clock::now().time_since_epoch()).count()) / 1000.0;
 }
